requirements.txt
asgiref==3.6.0
cfgv==3.3.1
distlib==0.3.6
Django==4.1.7
filelock==3.9.0
identify==2.5.18
nodeenv==1.7.0
platformdirs==3.0.0
pre-commit==3.0.4
PyYAML==6.0
sqlparse==0.4.3
tzdata==2022.7
virtualenv==20.19.0
djangorestframework-oauth==1.1.0
